+++ lib/CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Feat
-- add containerProps to Chart props
+- add `containerProps` and `options` to Chart props
+- add paneIndex series prop
 
 ## [0.2.1] - 2025-04-08
 ### Fix